+++ src/main/scala/parsley/Parsley.scala
@@ -1,9 +1,7 @@
 package parsley
 
-import parsley.deepembedding.ContOps._
-import parsley.deepembedding.DeepToken._
-import parsley.instructions._
-import parsley.deepembedding.{Parsley => _, _}
+import parsley.internal.instructions._
+import parsley.internal.deepembedding
 
 import scala.annotation.tailrec
 import scala.collection.mutable
@@ -85,7 +83,7 @@ object Parsley
           * @return A new parser which parses `pf`, then `px` then applies the value returned by `px` to the function
           *         returned by `pf`
           */
-        def <*>[B, C](px: =>Parsley[B])(implicit ev: P <:< Parsley[B=>C]): Parsley[C] = new Parsley(new DeepEmbedding.<*>[B, C](ev(p).internal, px.internal))
+        def <*>[B, C](px: =>Parsley[B])(implicit ev: P <:< Parsley[B=>C]): Parsley[C] = new Parsley(new deepembedding.<*>[B, C](ev(p).internal, px.internal))
         /**
           * This is the traditional Monadic binding operator for parsers. When the invokee produces a value, the function
           * `f` is used to produce a new parser that continued the computation.
@@ -95,7 +93,7 @@ object Parsley
           * @param f A function that produces the next parser
           * @return The parser produces from the application of `f` on the result of the last parser
           */
-        def flatMap[B](f: A => Parsley[B]): Parsley[B] = new Parsley(new DeepEmbedding.>>=(p.internal, f.andThen(_.internal)))
+        def flatMap[B](f: A => Parsley[B]): Parsley[B] = new Parsley(new deepembedding.>>=(p.internal, f.andThen(_.internal)))
         /**This combinator is an alias for `flatMap(identity)`.*/
         def flatten[B](implicit ev: A <:< Parsley[B]): Parsley[B] = flatMap[B](ev)
 
@@ -114,7 +112,7 @@ object Parsley
           * @return The value produced by the invokee if it was successful, or if it failed without consuming input, the
           *         possible result of parsing q.
           */
-        def <|>[B >: A](q: =>Parsley[B]): Parsley[B] = new Parsley(new DeepEmbedding.<|>(p.internal, q.internal))
+        def <|>[B >: A](q: =>Parsley[B]): Parsley[B] = new Parsley(new deepembedding.<|>(p.internal, q.internal))
         /**This combinator is defined as `p <|> pure(x)`. It is pure syntactic sugar.*/
         def </>[B >: A](x: B): Parsley[B] = this <|> pure(x)
         /**This combinator is an alias for `<|>`.*/
@@ -129,14 +127,14 @@ object Parsley
           * @param q The parser whose result should be returned
           * @return A new parser which first parses `p`, then `q` and returns the result of `q`
           */
-        def *>[A_ >: A, B](q: =>Parsley[B]): Parsley[B] = new Parsley(new DeepEmbedding.*>[A_, B](p.internal, q.internal))
+        def *>[A_ >: A, B](q: =>Parsley[B]): Parsley[B] = new Parsley(new deepembedding.*>[A_, B](p.internal, q.internal))
         /**
           * This is the parser that corresponds to a more optimal version of `p.map(x => _ => x) <*> q`. It performs
           * the parse action of both parsers, in order, but discards the result of the second parser.
           * @param q The parser who should be executed but then discarded
           * @return A new parser which first parses `p`, then `q` and returns the result of the `p`
           */
-        def <*[B](q: =>Parsley[B]): Parsley[A] = new Parsley(new DeepEmbedding.<*(p.internal, q.internal))
+        def <*[B](q: =>Parsley[B]): Parsley[A] = new Parsley(new deepembedding.<*(p.internal, q.internal))
         /**
           * This is the parser that corresponds to `p *> pure(x)` or a more optimal version of `p.map(_ => x)`.
           * It performs the parse action of the invokee but discards its result and then results the value `x` instead
@@ -157,7 +155,7 @@ object Parsley
           * @param pred The predicate that is tested against the parser result
           * @return The result of the invokee if it passes the predicate
           */
-        def filter(pred: A => Boolean): Parsley[A] = new Parsley(new DeepEmbedding.Ensure(p.internal, pred))
+        def filter(pred: A => Boolean): Parsley[A] = new Parsley(new deepembedding.Ensure(p.internal, pred))
         /** Filter the value of a parser; if the value returned by the parser does not match the predicate `pred` then the
          * filter succeeded, otherwise the parser fails with an empty error
          * @param pred The predicate that is tested against the parser result
@@ -171,7 +169,7 @@ object Parsley
           * @param msg The message used for the error if the input failed the check
           * @return The result of the invokee if it passes the predicate
           */
-        def guard(pred: A => Boolean, msg: String): Parsley[A] = new Parsley(new DeepEmbedding.Guard(p.internal, pred, msg))
+        def guard(pred: A => Boolean, msg: String): Parsley[A] = new Parsley(new deepembedding.Guard(p.internal, pred, msg))
         /** Similar to `filter`, except the error message desired is also provided. This allows you to name the message
           * itself. The message is provided as a generator, which allows the user to avoid otherwise expensive
           * computation.
@@ -179,7 +177,7 @@ object Parsley
           * @param msggen Generator function for error message, generating a message based on the result of the parser
           * @return The result of the invokee if it passes the predicate
           */
-        def guard(pred: A => Boolean, msggen: A => String): Parsley[A] = new Parsley(new DeepEmbedding.FastGuard(p.internal, pred, msggen))
+        def guard(pred: A => Boolean, msggen: A => String): Parsley[A] = new Parsley(new deepembedding.FastGuard(p.internal, pred, msggen))
         /** Similar to `filterNot`, except the error message desired is also provided. This allows you to name the message
          * itself.
          * @param pred The predicate that is tested against the parser result
@@ -200,7 +198,7 @@ object Parsley
         /**Alias for guard combinator, taking a dynamic message generator.*/
         def >?>(pred: A => Boolean, msggen: A => String): Parsley[A] = guard(pred, msggen)
         /**Sets the expected message for a parser. If the parser fails then `expected msg` will added to the error*/
-        def ?(msg: String): Parsley[A] = new Parsley(new DeepEmbedding.ErrorRelabel(p.internal, msg))
+        def ?(msg: String): Parsley[A] = new Parsley(new deepembedding.ErrorRelabel(p.internal, msg))
         /**Hides the "expected" error message for a parser.*/
         def hide: Parsley[A] = ?("")
         /** Same as `fail`, except allows for a message generated from the result of the failed parser. In essence, this
@@ -208,14 +206,14 @@ object Parsley
           * @param msggen The generator function for error message, creating a message based on the result of invokee
           * @return A parser that fails if it succeeds, with the given generator used to produce the error message
           */
-        def !(msggen: A => String): Parsley[Nothing] = new Parsley(new DeepEmbedding.FastFail(p.internal, msggen))
+        def !(msggen: A => String): Parsley[Nothing] = new Parsley(new deepembedding.FastFail(p.internal, msggen))
         /** Same as `unexpected`, except allows for a message generated from the result of the failed parser. In essence,
           * this is equivalent to `p >>= (x => unexpected(x))` but requires no expensive computations from the use of
           * `>>=`
           * @param msggen The generator function for error message, creating a message based on the result of invokee
           * @return A parser that fails if it succeeds, with the given generator used to produce an unexpected message
           */
-        def unexpected(msggen: A => String): Parsley[Nothing] = new Parsley(new DeepEmbedding.FastUnexpected(p.internal, msggen))
+        def unexpected(msggen: A => String): Parsley[Nothing] = new Parsley(new deepembedding.FastUnexpected(p.internal, msggen))
         /** Transforms this parser into a subroutine; instead of inlining this parser into every use-site (performing
           * optimisations and repeatedly generating code), produces a subroutine-like parser which is jumped to when
           * required. This will introduce runtime overhead, but it is fairly cheap and speeds up the compilation
@@ -223,7 +221,7 @@ object Parsley
           * @return The same parser, but wrapped in a subroutine call
           */
         @deprecated("This functionality is now automatically provided by default")
-        def unary_+ : Parsley[A] = new Parsley(new DeepEmbedding.Subroutine(p.internal))
+        def unary_+ : Parsley[A] = new Parsley(new deepembedding.Subroutine(p.internal))
         /**
           * Using this method enables debugging functionality for this parser. When it is entered a snapshot is taken and
           * presented on exit. It will signify when a parser is entered and exited as well. Use the break parameter to halt
@@ -231,7 +229,7 @@ object Parsley
           * @param name The name to be assigned to this parser
           * @param break The breakpoint properties of this parser, defaults to NoBreak
           */
-        def debug[A_ >: A](name: String, break: Breakpoint = NoBreak): Parsley[A_] = new Parsley(new DeepEmbedding.Debug[A_](p.internal, name, break))
+        def debug[A_ >: A](name: String, break: Breakpoint = NoBreak): Parsley[A_] = new Parsley(new deepembedding.Debug[A_](p.internal, name, break))
         /**
           * A fold for a parser: `p.foldRight(k)(f)` will try executing `p` many times until it fails, combining the
           * results with right-associative application of `f` with a `k` at the right-most position
@@ -271,7 +269,7 @@ object Parsley
           * @param b The parser that yields the condition value
           * @return The result of either `p` or `q` depending on the return value of the invokee
           */
-        def ?:(b: =>Parsley[Boolean]): Parsley[A] = new Parsley(new DeepEmbedding.Ternary(b.internal, p.internal, q.internal))
+        def ?:(b: =>Parsley[Boolean]): Parsley[A] = new Parsley(new deepembedding.Ternary(b.internal, p.internal, q.internal))
     }
 
     /** This is the traditional applicative pure function (or monadic return) for parsers. It consumes no input and
@@ -280,7 +278,7 @@ object Parsley
       * @param x The value to be returned from the parser
       * @return A parser which consumes nothing and returns `x`
       */
-    def pure[A](x: A): Parsley[A] = new Parsley(new DeepEmbedding.Pure(x))
+    def pure[A](x: A): Parsley[A] = new Parsley(new deepembedding.Pure(x))
 
     /** `lift1(f, p)` is an alias for `p.map(f)`. It is provided for symmetry with lift2 and lift3 */
     def lift1[A, B](f: A => B, p: =>Parsley[A]): Parsley[B] = p.map(f)
@@ -293,7 +291,7 @@ object Parsley
       * @param q The second parser to parse
       * @return `f(x, y)` where `x` is the result of `p` and `y` is the result of `q`.
       */
-    def lift2[A, B, C](f: (A, B) => C, p: =>Parsley[A], q: =>Parsley[B]): Parsley[C] = new Parsley(new DeepEmbedding.Lift2(f, p.internal, q.internal))
+    def lift2[A, B, C](f: (A, B) => C, p: =>Parsley[A], q: =>Parsley[B]): Parsley[C] = new Parsley(new deepembedding.Lift2(f, p.internal, q.internal))
     /** Traditionally, `lift2` is defined as `lift3(f, p, q, r) = p.map(f) <*> q <*> r`. However, `f` is actually uncurried,
       * so it's actually more exactly defined as; read `p` and then read `q` and then read 'r' then provide their results
       * to function `f`. This is designed to bring higher performance to any curried operations that are not themselves
@@ -304,7 +302,7 @@ object Parsley
       * @param r The third parser to parse
       * @return `f(x, y, z)` where `x` is the result of `p`, `y` is the result of `q` and `z` is the result of `r`.
       */
-    def lift3[A, B, C, D](f: (A, B, C) => D, p: =>Parsley[A], q: =>Parsley[B], r: =>Parsley[C]): Parsley[D] = new Parsley(new DeepEmbedding.Lift3(f, p.internal, q.internal, r.internal))
+    def lift3[A, B, C, D](f: (A, B, C) => D, p: =>Parsley[A], q: =>Parsley[B], r: =>Parsley[C]): Parsley[D] = new Parsley(new deepembedding.Lift3(f, p.internal, q.internal, r.internal))
 
     /** This is one of the core operations of a selective functor. It will conditionally execute one of `p` and `q`
       * depending on the result from `b`. This can be used to implement conditional choice within a parser without
@@ -336,29 +334,29 @@ object Parsley
       * @param p The parser to run
       * @return The result of `p`, or if `p` failed ensures the parser state was as it was on entry.
       */
-    def attempt[A](p: =>Parsley[A]): Parsley[A] = new Parsley(new DeepEmbedding.Attempt(p.internal))
+    def attempt[A](p: =>Parsley[A]): Parsley[A] = new Parsley(new deepembedding.Attempt(p.internal))
     /** Parses `p` without consuming any input. If `p` fails and consumes input then so does `lookAhead(p)`. Combine with
       * `attempt` if this is undesirable.
       * @param p The parser to look ahead at
       * @return The result of the lookahead
       */
-    def lookAhead[A](p: =>Parsley[A]): Parsley[A] = new Parsley(new DeepEmbedding.Look(p.internal))
+    def lookAhead[A](p: =>Parsley[A]): Parsley[A] = new Parsley(new deepembedding.Look(p.internal))
     /**Alias for `p ? msg`.*/
     def label[A](p: Parsley[A], msg: String): Parsley[A] = p ? msg
     /** The `fail(msg)` parser consumes no input and fails with `msg` as the error message */
-    def fail(msg: String): Parsley[Nothing] = new Parsley(new DeepEmbedding.Fail(msg))
+    def fail(msg: String): Parsley[Nothing] = new Parsley(new deepembedding.Fail(msg))
     /** The `empty` parser consumes no input and fails softly (that is to say, no error message) */
-    val empty: Parsley[Nothing] = new Parsley(new DeepEmbedding.Empty)
+    val empty: Parsley[Nothing] = new Parsley(new deepembedding.Empty)
     /** The `unexpected(msg)` parser consumes no input and fails with `msg` as an unexpected error */
-    def unexpected(msg: String): Parsley[Nothing] = new Parsley(new DeepEmbedding.Unexpected(msg))
+    def unexpected(msg: String): Parsley[Nothing] = new Parsley(new deepembedding.Unexpected(msg))
     /** Returns `()`. Defined as `pure(())` but aliased for sugar*/
     val unit: Parsley[Unit] = pure(())
     /** converts a parser's result to () */
     def void(p: Parsley[_]): Parsley[Unit] = p *> unit
     /** `many(p)` executes the parser `p` zero or more times. Returns a list of the returned values of `p`. */
-    def many[A](p: =>Parsley[A]): Parsley[List[A]] = new Parsley(new DeepEmbedding.Many(p.internal))
+    def many[A](p: =>Parsley[A]): Parsley[List[A]] = new Parsley(new deepembedding.Many(p.internal))
     /** `skipMany(p)` executes the parser `p` zero or more times and ignores the results. Returns `()` */
-    def skipMany[A](p: =>Parsley[A]): Parsley[Unit] = new Parsley(new DeepEmbedding.SkipMany(p.internal))
+    def skipMany[A](p: =>Parsley[A]): Parsley[Unit] = new Parsley(new deepembedding.SkipMany(p.internal))
     /**
       * Evaluate each of the parsers in `ps` sequentially from left to right, collecting the results.
       * @param ps Parsers to be sequenced
@@ -382,12 +380,12 @@ object Parsley
       * This parser consumes no input and returns the current line number reached in the input stream
       * @return The line number the parser is currently at
       */
-    val line: Parsley[Int] = new Parsley(DeepEmbedding.Line)
+    val line: Parsley[Int] = new Parsley(deepembedding.Line)
     /**
       * This parser consumes no input and returns the current column number reached in the input stream
       * @return The column number the parser is currently at
       */
-    val col: Parsley[Int] = new Parsley(DeepEmbedding.Col)
+    val col: Parsley[Int] = new Parsley(deepembedding.Col)
     /**
       * This parser consumes no input and returns the current position reached in the input stream
       * @return Tuple of line and column number that the parser has reached
@@ -400,7 +398,7 @@ object Parsley
       * @tparam S The type of the value in register `v` (this will result in a runtime type-check)
       * @return The value stored in register `v` of type `S`
       */
-    def get[S](v: Var)(implicit ev: S =!= Nothing): Parsley[S] = new Parsley(new DeepEmbedding.Get(v))
+    def get[S](v: Var)(implicit ev: S =!= Nothing): Parsley[S] = new Parsley(new deepembedding.Get(v))
     /**
       * Consumes no input and returns the value stored in one of the parser registers after applying a function.
       * @note There are only 4 registers at present.
@@ -434,7 +432,7 @@ object Parsley
       * @param v The index of the register to place the value in
       * @param p The parser to derive the value from
       */
-    def put[S](v: Var, p: =>Parsley[S]): Parsley[Unit] = new Parsley(new DeepEmbedding.Put(v, p.internal))
+    def put[S](v: Var, p: =>Parsley[S]): Parsley[Unit] = new Parsley(new deepembedding.Put(v, p.internal))
     /**
       * Modifies the value contained in register `v` using function `f`. It is left to the users responsibility to
       * ensure the types line up. There is no compile-time type checking enforced!
@@ -443,7 +441,7 @@ object Parsley
       * @param f The function used to modify the register
       * @tparam S The type of value currently assumed to be in the register
       */
-    def modify[S](v: Var, f: S => S): Parsley[Unit] = new Parsley(new DeepEmbedding.Modify(v, f))
+    def modify[S](v: Var, f: S => S): Parsley[Unit] = new Parsley(new deepembedding.Modify(v, f))
     /**
       * For the duration of parser `p` the state stored in register `v` is instead set to `x`. The change is undone
       * after `p` has finished.
@@ -463,7 +461,7 @@ object Parsley
       * @param q The parser to execute with the adjusted state
       * @return The parser that performs `q` with the modified state
       */
-    def local[R, A](v: Var, p: =>Parsley[R], q: =>Parsley[A]): Parsley[A] = new Parsley(new DeepEmbedding.Local(v, p.internal, q.internal))
+    def local[R, A](v: Var, p: =>Parsley[R], q: =>Parsley[A]): Parsley[A] = new Parsley(new deepembedding.Local(v, p.internal, q.internal))
     /**
       * For the duration of parser `p` the state stored in register `v` is instead modified with `f`. The change is undone
       * after `p` has finished.
