+++ src/main/scala/parsley/internal/instructions/package.scala
@@ -1,4 +1,4 @@
-package parsley
+package parsley.internal
 
 import scala.annotation.tailrec
 
@@ -9,32 +9,66 @@ package object instructions
     private [instructions] case object Recover extends Status
     private [instructions] case object Failed extends Status
 
-    private [parsley] abstract class Instr
-    {
+    private [internal] abstract class Instr {
         def apply(ctx: Context): Unit
         // Instructions should override this if they have mutable state inside!
         def copy: Instr = this
     }
 
-    private [parsley] trait Stateful
+    private [internal] trait Stateful
 
-    private [parsley] abstract class JumpInstr extends Instr
-    {
+    private [internal] abstract class JumpInstr extends Instr {
         var label: Int
     }
 
     // It's 2018 and Labels are making a come-back, along with 2 pass assembly
-    private [parsley] final class Label(val i: Int) extends Instr
-    {
+    private [internal] final class Label(val i: Int) extends Instr {
         def apply(ctx: Context): Unit = throw new Exception("Cannot execute label")
+    }
+
+    final private [parsley] def pretty(instrs: Array[Instr]): String = {
+        val n = instrs.length
+        val digits = if (n != 0) Math.log10(n).toInt + 1 else 0
+        instrs.zipWithIndex.map {
+            case (instr, idx) =>
+                val paddedIdx = {
+                    val str = idx.toString
+                    " " * (digits - str.length) + str
+                }
+                val paddedHex = {
+                    val str = instr.##.toHexString
+                    "0" * (8 - str.length) + str
+                }
+                s"$paddedIdx [$paddedHex]: $instr"
+        }.mkString(";\n")
+    }
+
+    final private [internal] def stateSafeCopy(instrs: Array[Instr], pindices: Array[Int]): Array[Instr] = {
+        val nstateful = pindices.length
+        if (nstateful != 0) {
+            val instrs_ = instrs.clone
+            for (i <- 0 until nstateful) {
+                val j = pindices(i)
+                instrs_(j) = instrs(j).copy
+            }
+            instrs_
+        }
+        else instrs
+    }
+
+    final private [internal] def statefulIndices(instrs: Array[Instr]): Array[Int] = {
+        val buff = new ResizableArray[Int]()
+        for (i <- 0 until instrs.length) {
+            if (instrs(i).isInstanceOf[Stateful]) buff += i
+        }
+        buff.toArray
     }
 
     // This stack class is designed to be ultra-fast: no virtual function calls
     // It will crash with NullPointerException if you try and use head or tail of empty stack
     // But that is illegal anyway
     private [instructions] final class Stack[A](var head: A, val tail: Stack[A])
-    private [instructions] object Stack
-    {
+    private [instructions] object Stack {
         def empty[A]: Stack[A] = null
         @inline def isEmpty(s: Stack[_]): Boolean = s == null
         @tailrec def drop[A](s: Stack[A], n: Int): Stack[A] = if (n > 0 && !isEmpty(s)) drop(s.tail, n - 1) else s
@@ -46,16 +80,13 @@ package object instructions
     // Designed to replace the operational stack
     // Since elements are of type Any, this serves as a optimised implementation
     // Its success may result in the deprecation of the Stack class in favour of a generic version of this!
-    private [instructions] final class ArrayStack[A](initialSize: Int = 8)
-    {
+    private [instructions] final class ArrayStack[A](initialSize: Int = 8) {
         private [this] var array: Array[Any] = new Array(initialSize)
         private [this] var sp = -1
 
-        def push(x: A): Unit =
-        {
+        def push(x: A): Unit = {
             sp += 1
-            if (array.length == sp)
-            {
+            if (array.length == sp) {
                 val newArray: Array[Any] = new Array(sp * 2)
                 java.lang.System.arraycopy(array, 0, newArray, 0, sp)
                 array = newArray
@@ -64,15 +95,13 @@ package object instructions
         }
 
         def exchange(x: A): Unit = array(sp) = x
-        def peekAndExchange(x: A): Any =
-        {
+        def peekAndExchange(x: A): Any = {
             val y = array(sp)
             array(sp) = x
             y
         }
         def pop_(): Unit = sp -= 1
-        def upop(): Any =
-        {
+        def upop(): Any = {
             val x = array(sp)
             sp -= 1
             x
@@ -91,18 +120,15 @@ package object instructions
         def size: Int = usize + 1
         def isEmpty: Boolean = sp == -1
         def mkString(sep: String): String = array.take(sp+1).reverse.mkString(sep)
-        def clear(): Unit =
-        {
+        def clear(): Unit = {
             sp = -1
             var i = array.length-1
-            while (i >= 0)
-            {
+            while (i >= 0) {
                 array(i) = null
                 i -= 1
             }
         }
-        override def clone: ArrayStack[A] =
-        {
+        override def clone: ArrayStack[A] = {
             val me = new ArrayStack[A](array.length)
             for (i <- 0 to sp) me.push(array(i).asInstanceOf[A])
             me
