+++ src/main/scala/parsley/internal/package.scala
@@ -1,0 +1,5 @@
+package parsley
+
+package object internal {
+    private [internal] type UnsafeOption[A >: Null] = A
+}
