+++ src/test/scala/parsley/internal/InternalTests.scala
@@ -1,0 +1,17 @@
+package parsley.internal
+
+import parsley.{ParsleyTest, Success}
+import parsley.Parsley._
+import parsley.Char.{char, satisfy}
+import parsley.Implicits.charLift
+
+import scala.language.implicitConversions
+
+class InternalTests extends ParsleyTest {
+    "subroutines" should "function correctly and be picked up" in {
+        val p = satisfy(_ => true) *> satisfy(_ => true) *> satisfy(_ => true)
+        val q = 'a' *> p <* 'b' <* p <* 'c'
+        q.internal.instrs.last should be (instructions.Return)
+        q.runParser("a123b123c") should be (Success('3'))
+    }
+}
